backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.32.1
python-multipart==0.0.22
pymupdf==1.25.1
openai==1.59.3
nbformat==5.10.4
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.4
